Convert audio effects to mono channel.
We that for some reason, a lot of SS13's audio effects sound *bad* in stereo mode. BYOND seems to convert them to Mono internally while playing them which circumvents this.
This merges the audio tracks into one, saving some space and making them sound correct on Godot.

Also, OpenAL doesn't like playing Stereo audio positionally, and I'm not gonna write a channel merger in C# for this to work around this.