Add reagent spills
This is 50% design 50% code review. There's some ugly stuff in there that I wasn't sure how to implement better.
Essentially there's a separate entity for each tile the spill is on so the puddles function like a solution container. The only way to remove from it is with a mop but anything can add to it.
The coloring uses the chemical colors (I didn't add all of the sprites from Eris as the remaining ones had kind of a messy layout) unless overridden (currently for vomit). The opacity is based on how much there is with a set amount before it spills to neighboring tiles.
If the solution amount is low (currently it's set to 2 or less units) it will delete itself eventually.

I also touched:
Chemical colors (added some more for testing)
Chemical color calculator (not sure if this was working?)

Still TODO:
Slips
Tracks