Add puddles / reagent spills
Take two.

Puddles are an entity that uses the snapgrid so there's a max of 1 per tile. Mops and buckets can transfer this entity's solution elsewhere (and the entity is deleted if it's empty).
Puddles will also overflow once they get too full / evaporate if minimal solution for a long time (potentially change by design?)
It's also possible to override the re-colouring (e.g. if you want a generic colour like vomit), otherwise it'll use the regular chemical re-colouring (which didn't seem to work before?)
Some other stuff also seemed to throw since I started this yonks ago so I updated that as well.

The sprite to be used will also be randomly selected unless overwritten with the spillhelper (e.g. if someone gets slashed you could force a specific type of sprite).