Alerts rework
Reworked some alert images so they aren't so rough around the edges, as well as adding an original drawing for being pulled (stun icon has a different route and foulder but is still the turned off stun baton, someone should draw a better one and I am not up to the task). There are minor description changes so they sounds either more resonable or more informational.