Add CandleComponent, CandleSystem and CandleStateVisualizer
Created CandleComponent, CandleSystem and CandleStateVisualizer

These are heavily based off of MatchstickComponent, MatchstickSystem and BurnStateVisualizer, i've added a bunch of stuff for tracking Candle duration 'wax' and up to 4 candle states.

When a candle component is initialized it will also slightly randomize the burn duration so that no 2 candles burn the exact same amount of time.