Let items on conveyor belts round corners
Basic idea: if the item is not in the middle of the conveyor in the
direction that the conveyor is running, move the item towards the
middle. Otherwise, move the item along the direction. This lets
conveyors pick up items that are not perfectly aligned in the middle,
and also makes corner cuts work.

We do this by computing the projection of 'itemRelative' on 'direction',
yielding a vector 'p' in the direction of 'direction'. We also compute
the rejection 'r'. If the magnitude of 'r' is not (near) zero, then the
item is not on the centerline.