Add "already full" message when refilling welder
Due to a logic error, refilling a full welder on a fuel tank shows the
"fuel tank is empty" message. This change adds an appropriate message to
show that the welder is already full.

While here, adjust the entity name of the fuel tank (and other entities
in the same file) to uncapitalize the first letter to be consistent with
all the other entity names.