Emagged Chem Dispenser has Water and Welding Fuel
This addition makes Emagging a chem Dispenser more useful. As it is, the only item in the Emagged Chem Dispenser that is truly useful is the Epinephrine as the other item's aren't very useful to a traitor chemist. The addition of water and Welding Fuel makes them more useful to a chemist.