SmartFridge now contains chemistry equipment
Added 3 boxes of bottles, 3 boxes of pill canisters, and 3 boxes of beakers to the Smart Fridge. Eventually I want chemists to be able to dump bottles and pills into the fridge so doctors can come over and dispense them, but I think this is a good change to make it useful for now.