Config presets system.
This moves a couple OverrideDefaults behind #if checks into dedicated TOML files in the resources. A couple CVars can now turn these overrides off too so they're easier to disable.

A new system is also provided to allow config-defined presets to be loaded. Idea here is that some Wizard's Den specific config values can be put in these. They'd be tracked in repo but need to be manually enabled in the server config.

Also, game.lobbyenabled is now default ON, but put behind the new development TOML file. This means that downloading a zip version of the server will have lobby enabled by default, but for development there's no change.