Fix S.T.Json holding integration instances live for long.
System.Text.Json introduced a cache in .NET 7 to deal with people not caching their JsonSerializerOptions. This now means that it keeps stuff live without any sort of CWT<,>, so integration instances stay alive for way too long, reducing performance and bloating memory usage.

Use weak references in the JsonConverters so that this doesn't happen.