Increase handcuffs' attack speed
PR #12838 made it possible to cuff in harm mode, but it sometimes
feels like it doesn't work. That's because of the attack cooldown.

When you take the cuffs out of your bag, you have to wait a bit before
you can cuff in harm mode. Increasing the attack speed reduces that
time enough to be insignificant.