Make paused poweredlights enabled
Even if they have no power in range. I think this is preferable over keeping them in darkness? Power stuff shouldn't be running when the sim is paused so atm they just check for whether they're paused; doesn't seem to cause issues with mapinitTM. Worst case we just revert and wait for power to be fixed.
Alternatively you could have a command to toggle it but uhhh not sure.