Fix slowoncontacts mispredict
So I wasn't entirely sure if it was due to contact shenanigans on engine because predicted conveyors feel mostly fine.

As far as I can tell the main issue now is the RemCompDeferred issue. Ideally we would just be able to ensure these components but that causes a mispredict between the component removal so adding it to prototypes seemed like the lesser evil at the moment.