Avoid NPCs getting stuck around tables
There's a shortcut if NPCs can move directly to a target but this was causing issues due to table collision masks so this will avoid it.

I also avoided some unintentional node pruning e.g. if the NPC needs to go backwards to get closer to the target it would accidentally prune nodes sometimes.