Update SpawnPointSystem.cs
It would be more logical to start by trying to throw the player at a late join point and not throw him anywhere. For the most part, this is necessary in case of adding new professions, but not having the desire to add a marker for a new profession to each card.