Remove bounds check for FTL
It's annoying and leaks meta-knowledge of nukies and stuff and in hindsight is a bad idea.

Cargo still have to drive in but they can just FTL from next to the station which arrivals and emergency do anyway. If we want to stop FTLing away for traitors it should be like a tractor beam or smth.