Add instant destruction thresholds to everything
When a nuke goes off the audio floods sources so much that the console is flooded.

In the distant future we could potentially clump these together though we need a lot of audio work before that point.

This probably means less debris although I think not hitting the audio source limit is more important.