Turn culture invariant off by default
This commit turns the culture invariant off by default. When the culture
invariant is on, no localization is allowed, and setting the locale to
the default value of en-US fails. I'm not sure why this change is
needed, because according to the dotnet docs it should be off by default
anyways, but this was causing breakage on linux. This commit is
confirmed to fix it on my system, but might need testing on windows to
make sure it doesnt break things.

Fixes #10305