Predatory pet for HoS
Adds your own pet for the station's main defender. Since it is a spider, it is a predatory and poisonous insect. If it notices other spiders, rats or other aggressive fauna on its territory (except Dragons, none of the beasts fight Dragons), it will definitely attack.
I had to make a separate faction, because with the NT faction, Shiva would have paranormal powers in search of Syndicates. Which of course would be fun, but every shift we'd have to see a HoS spider walking on a leash with her, who just decided to walk the corridors of the station with her.