Triple strength of THC
Smokable THC like joints and blunts do not show the status effect
until it has run a sufficient time. With the previous SeeingRainbows
time of 5, the effect would end shortly after it started because
smokable items do not inhale enough solution for it to continue.

This change introduces a tripling of the time added to the
SeeingRainbows effect in order to counteract the lack of visible
effect from smokables.

Now, the effect will trigger within a few seconds and last a little
over a minute after the blunt has burned out.