Improve HandLabeler and AgentID UI
Previously, both UIs required you to hit enter in order for the event
to trigger. Now, it sends the message when focus is lost, allowing for
both items to update even if the player doesn't hit enter, which tends
to be the more natural thing to do. It works even if you directly close
the window, as the FocusExit event fires before that one.

The HandLabeler's ActivateInWorld handler was replaced with the more
suitable ActivatableUI component used in many other devices.