Limit postgres database concurrency
We noticed that round end/start causes huge spikes of idle connections on the DB server, causing other servers to report connection limit errors.

While this is definitely an issue that should be fixed in other game code *somewhere* (so we avoid spamming connections in the first place) I'd rather have a safety measure either way.