Fix bugs when updating paper UI
Paper UI update events were being broadcast to all players. This led to
some bugs:

If player A is writing a paper and player B reads it, player A's editing
interface disappears and they lose the text they input so far.

If player A is reading a paper and player B starts writing it, player A
will get an editing interface and be able to write paper without a pen.

Now, the editing interface is available only to players who actually
started writing a paper, and the editing interface is only updated when
the text changes -- this removes the possibility of losing
partially-input text in all but one (rare) case:

If player A starts writing a paper but player B starts (and finishes)
writing the same paper, then player A's in-progress text is lost. Both
players will see player B's updated text.

Short of implementing a realtime, collaborative document or changing how
text is input, I think we'll have to live with it.

Bonus: Removes usage of some obsolete methods in PaperSystem.cs