Add station anchors
This modifies the hard-coded shuttle linear and angular damping
coefficient to be a data field on the shuttle component. This allows a
shuttle to override the grid's linear and angular damping. If you set
this to a sufficiently large number, these behave like station anchors.

Since the default shuttle damping is the old hard-coded value, this
change does not affect any existing maps. Mappers must update their maps
to enable station anchors.