Align filled gas tanks with intended capacity
The gas tanks changed in PR #17106 were only the empty ones.
None of the filled tanks were altered, so the change was functioning
only if you took a tank that spawns empty, filled it, and used it.

All filled tanks have had their moles set to a value to reflect 1000 kPa,
as if someone had filled them recently from a canister.

Formulas have been left behind in the comments to teach others in the
future, in the event someone wants to change these values again.

The nitrous oxide tank was altered as well in order to have a more
sensible output pressure for its contained gasses. I had to do some
guesswork and testing on how much N2O to put into the tank in order to
get someone to fall asleep shortly after engaging the tank.

Removed superfluous description from plasma tank which was already
inherited from its parent.