(Probably) fix random integration test failures.
I noticed that the sporadic test failures all involved RobustSynchronizationContext. This should in and of itself be impossible: integration tests are *supposed* to be configured with the database running fully synchronous.

The actual source of the bug is the admin log manager deferring writes via Task.Run(). This caused multiple requests to hit the concurrency semaphore at the same time, which caused the main integration testing logic to wait on the rest of the thread pool, introducing divergence. From there on it's no telling what goes wrong with the integration testing system, everything's timed wrong and the database is potentially written to across instances.

The funny thing is, that Task.Run() in admin logs isn't even necessary anymore. Nowadays, the DB manager itself forces a Task.Run() regardless (except during integration tests). The simplest fix is to just remove it from admin logs entirely.

The rest of the commit is just a ton of guard code around the database that I used to debug this. It also serves to make sure this never happens again, enforcing that database queries are always returned synchronously during integration tests. This bug became remarkably easy to reproduce as soon as basic .IsCompleted checks were in place to spot the concurrency issues.

I also chased a bit of a red herring wrt admin logs' usage of IAsyncEnumerable<T>. I forced it to not load the query results incrementally (so it's checked synchronous), figuring tests wouldn't have data sizes for this to matter. I realize now while writing this commit message that such a change may mask potential deadlock bugs otherwise caught by integration tests, so I will try to remember to revise this with a more thorough fix later.