Offset shot projectiles by gun barrel length
Add a new barrel length field to gun components, which determines how
far from the center of the player the bullet is spawned.

This offset places laser gun effects offset from the gun as if it came
from the end of the barrel, instead of starting the beam right in the
middle of the player.

This offset also fixes the problem of players on vehicles shooting
themselves, without making vehicles completely transparent to
projectiles.