firearm cycling fixed
Firearms have a Cycled property.
If not AutoCycle, then Cycled will be set to false upon firing (or loading ammo from the BallisticAmmoProvider component to be exact, which happens when firing).
Further ammo cannot be loaded when Cycled is set to false.
Manually cycling sets this to true, and doesn't eject a cartridge if not cycled yet.