Fix race condition causing some people being unable to connect.
This would happen if MsgPreferencesAndSettings arrived on the client before MsgPlayerList.

The LobbyCharacterPreviewPanel is for some reason always loaded by the UI system, meaning it was always hooked to the IClientPreferencesManager.OnServerDataLoaded event. If the preferences loaded before MsgPlayerList (which sets the client run level to "Connected"), it would try to spawn the preview human client entities before the entity system was initialized.

Fixed by moving the event subscription to LobbyState, so it's only listening when we're in the lobby (after the client is properly connected).