Jittering System Fix / Cleanup
This fixes a bug where the length of a jittering animation would be infinite (capped to 2 seconds) when the frequency was 0.
That can happen during the first jitter of a new component, causing an initial 2 second jitter that shouldn't be there.
Also removes the 2 second jittering limit cause it seems like a rather pointless limit and maybe band-aid fix for the aforementioned problem.