no anything
This reverts f391ff28 and implements an alternate messure where mice and other small animals can no longer infect people as zombies.
This is done through a component which if present cancels the check that would cause zombie components to be added on people that get infected due to a bite.
This still allows other special stuff that happens in that function that may affect already infected individuals.
This is a compromise between what's discussed in discord which would much rather see mice and other animals just die from the infection and people on github which would see no change happen.