Fix being unable to open chat channel drop-down.
So this broke because ChannelSelectorButton enabled EnableAllKeybinds (?) as a workaround for some weird popup closing issue (??) and then had to implement a bad hack (???) to ignore the button toggling twice due to multiple keybinds being bound against UIClick *and* Use. Which then broke because Emo added another button to click for grid inv.

Just what?

Anyways just... get rid of all that nonsense???

With all that nonsense gone the popup doesn't close if you click on the button directly (it gets closed by the UI manager modal code, then gets re-opened by the button). We fix that by not re-opening the popup if you closed it in the current frame.