Raise cryo metabolism min temperature
Many people have been running into issues cooling bodies to the current
temperature. Make it a bit easier by raising the max temperature.

This is the temperature that the body has to change to in order to
metabolize the drug, so in practice cryo has to cool lower than this in
order to actually get the body to this temperature.