Rev flashes eventually time out if not re-applied
This is a pretty major change to how revs work that
might well badly destabilize the game mode so I'll
just preface this with the fact that the key numbers
for how this work are exposed to admin cvar, and
thus it can be turned on/off at will.

The revolutionary round mode is a classic SS13 mode
but it was raised at the latest maintenance meeting
related to it being a bit wonky on MRP, etc.

PR 23329 proposes making mildshields rarer to
try and prevent Sec from just rushing to mindshield
the station at a hint of revs. This doesn't fix any of the
basic problems with revs: it being highly disruptive
in on MRP, it causing stalemates and the only route Sec
have to winning against revs is to get mindshields,
which rely on, amongst other things, Cargo not being
a giant hole into space.

So, by default, this PR changes the rev gamemode by
making flashes only _temporarily_ convert somebody
to a revolutionary. How temporarily? By default:

AFTER FIFTEEN MINUTES: the rev has a 5% chance for
the flash to wear off.
AFTER SIXTEEN MINUTES: this increases to 10%
AFTER SEVENTEEN MINUTES: now it's 15%...
And so on.

This means that revs have to try and bring about the death
of the heads _reasonably quickly_ once they begin flashing
people.

HOWEVER, a head rev is also at liberty to re-flash all their
revs. So a small team of revs can be kept flashed and safe.

This is only a proposal (I'm only baby contributor pls no hate)
and I'm wanting to add more to this PR to round it out.