Fix tranquiliser shells not working (#23886)
* Fix tranquiliser darts not working

Fixes a bug introduced by 9f47079d02a872f0aaff84f4556b83d3c6f835ef which
was made to stop the fly-by fixture from triggering the system. This was
done by checking whether the fixture was hard. Apparently the
projectile's fixture is never hard as well. The change just makes it so
that check only succeeds when the fixture is a fly-by fixture.

* Remove something that I think is redundant

* Remove random using directive that somehow appeared.

* Address Review

* Adress Review 2

* Put the appropriate fixture ids