Fix security camera sprite when viewed at an angle
The 8-directions is a vestige of /tg/station, where the extra
directions were used to place cameras on the left or right half of the
wall.

Setting it to 1 directions (sic) allows the game engine to rotate the
sprite for us, which is fine as the art currently has no perspective.