Bagels Initial Commit Pre-Feedback
Bagels are made by using a rolling pin on a dough slice to make a dough rope, then cooking the dough rope in a microwave for 5 seconds. There are two types: a normal bagel and a poppyseed bagel. The poppyseed bagel requires a poppy and a dough rope and has a small (5u) quantity of Bicaridine inside of it in addition to its nutriment.