Fix action container handling
I don't even know if this is the best way to fix it but this at least fixes it.

Pros:
- ActionsSystem is in charge of action addition / removal during state handling instead of containers doing it
Cons:
- If something removes an action from a container (atm this can happen if they get emptied for example) it may not update properly. In the future we could maybe add a new containertype for virtual entities or something.

I set the pvs budget to 5 and moved around and it appears fixed. Doesn't seem to cause any breaking changes.

I'd write a test but no idea how to repro the setup between client and server.