Records (#129)
* Implement records saving.

Add an additional column to the database that stores records as JSON.
Currently the serialization logic is not very robust so a future commit
should improve it so it can deal with missing and extra fields in the
database.

* Record editor gui

Implements a record editor tab for HumanoidProfileEditor. In addition,
this improves the robustness of the CharacterRecords deserializer.

* Add imperial units to record editor.

* Remove SecurityClearance from CharacterRecords

* Add up and down buttons to the record entry editor

This lets you select the order that records will appear in game.

* Add system for managing records in game

* Initial viewer UI

Add the record console UI, currently it does not display entries.

* Add entries to the record console

* Clarify record computers.

* Add records viewer to aghost

* Add admin modification ui for character records

* Improve record entry viewer

* Remove contact number record

* Character Records Guidebook Entry

* Change from DefaultWindow to FancyWindow

* Fix compile errors relating to rebase

* Add rich text to record viewer

We support all tags except the heading tag. This is because it looks
very ugly. In addition, we also add a view button to the editor GUI so
you can make sure your rich text works without going into a game.

* Integrate wanted status into security records

* Correctly subscribe to Bui events

* Add comments to records and fix a few minor issues

(cherry picked from commit 32af2d4ecdbb3167a6ee25b89559e5de6642a658)